+++ CHANGELOG.md
@@ -687,6 +687,8 @@ Compile via remote server
 -   Export parsing internals to `@promptbook/editable`
 -   Rename `sourceContent` -> `knowledgeSourceContent`
 -   Multiple functions to manipulate with `PipelineString`
+-   `book` notation supports values interpolation
+-   Make equivalent of `book` notation the `prompt` exported through `@promptbook/utils`
 -   Change admin email to `pavol@ptbk.io`
 
 ## In pre-release